admin_panel/requirements.txt
psycopg2==2.8.6
django==3.1.1
python-dotenv==0.14.0
gunicorn==20.0.4
django-apscheduler==0.5.2
djangorestframework==3.12.4
jinja2==2.11.3
pika==1.2.0